# Added STATIC_BMI2 for compile time detection of BMI2 on MSVC, when enabled various intrinsics are used