This adds STATIC_BMI2 define to compiler.h, which is enabled on MSVC when compiler target has __AVX2__ enabled

Like DYNAMIC_BMI2 but at compile time.

Replaces uses of BitScanForward/Reverse with _lzcnt_u64/_tzcnt_u64 and 32 variants for better code gen.

Other changes when STATIC_BMI2 is on:
BIT_getLowerBits uses _bzhi_u64 
BIT_getMiddleBits uses _bextr_u64

STATIC_BMI2 is only turned on with MSVC

MSVC is terrible at detecting patterns such as  _bextr_u64, clang is much better so maybe not necessary there.

In BIT_getLowerBits  Clang is smart enough to use _bzhi_u64 when -march=skylake is enabled(AVX2 support), MSVC fails to do so when /arch:AVX2 is on.