@Cyan4973 2020-07-28
Do you have some benchmark results ?
Or `perf` / IACA results ?