@Niadb 2020-07-28
Hi @Cyan4973 I used the fullbench program to test the changes in regards to performance.

I also used the VS profiler, it showed a few functions not being inlined, so I applied a force inline to a few hot functions in bitstream.h

The BitScanForward/Reverse changes were faster along with _bzhi_u64, but using _bextr_u64 in that context didn't work out, so I removed it. 

Here is the before and after of running fullbench.
Compiled with VS2019(latest), CPU is Ryzen 9 3900x, running Windows 10
AVX2 was enabled in build

Before:
*** Zstandard speed analyzer 1.4.5 64-bits, by Yann Collet (Jul 28 2020) ***
 Sample 10000000 bytes :
 1#compress                     :   520.1 MB/s  ( 3154550)
 2#decompress                   :  1727.9 MB/s  (10000000)
11#compressContinue             :   519.6 MB/s  ( 3154550)
12#compressContinue_extDict     :   512.6 MB/s  ( 3154887)
13#decompressContinue           :  1724.4 MB/s  (10000000)
31#decodeLiteralsBlock          :  2534.8 MB/s  (   37031)
32#decodeSeqHeaders             : 51772.6 MB/s  (      73)
41#compressStream               :   421.9 MB/s  ( 3154678)
42#decompressStream             :  1726.2 MB/s  (10000000)
43#compressStream_freshCCtx     :   421.9 MB/s  ( 3154678)
51#compress_generic, continue   :   422.2 MB/s  ( 3154678)
52#compress_generic, end        :   519.8 MB/s  ( 3154550)
61#compress_generic, -T2, contin:   421.3 MB/s  ( 3154678)
62#compress_generic, -T2, end   :   519.1 MB/s  ( 3154550)

After:
*** Zstandard speed analyzer 1.4.5 64-bits, by Yann Collet (Jul 28 2020) ***
 Sample 10000000 bytes :
 1#compress                     :   526.9 MB/s  ( 3154550)
 2#decompress                   :  1767.6 MB/s  (10000000)
11#compressContinue             :   527.1 MB/s  ( 3154550)
12#compressContinue_extDict     :   520.9 MB/s  ( 3154887)
13#decompressContinue           :  1767.5 MB/s  (10000000)
31#decodeLiteralsBlock          :  2522.1 MB/s  (   37031)
32#decodeSeqHeaders             : 50173.2 MB/s  (      73)
41#compressStream               :   419.7 MB/s  ( 3154678)
42#decompressStream             :  1755.4 MB/s  (10000000)
43#compressStream_freshCCtx     :   420.7 MB/s  ( 3154678)
51#compress_generic, continue   :   422.7 MB/s  ( 3154678)
52#compress_generic, end        :   524.6 MB/s  ( 3154550)
61#compress_generic, -T2, contin:   418.1 MB/s  ( 3154678)
62#compress_generic, -T2, end   :   526.1 MB/s  ( 3154550)