@Cyan4973 2020-07-28
Excellent, thank you @Niadb !
We are aware that Visual Studio performance is not as good as `gcc` or `clang`, so this patch is welcomed.
And yes, we also noticed that `bextr` is not an automatic win.

I'll have a more detailed look at this patch later today.