@Cyan4973 2020-07-30
I've been trying to replicate these results.
So far, results are slightly less good than anticipated from previous measurements, 
though broadly in line.

I'm unable to detect any consistent compression speed improvement. If there is, it's well within noise level.
On the decompression side though, there is a measurable speed improvement, just not a great one. The average is about ~3%, relatively consistent on synthetic data, but varies a lot on real datasets, with results in the [0-3%] range, generally lower.

Still, it's a net positive, so it qualifies,

Just for curiosity, I'm wondering if I do these speed tests correctly.
Apart from the usual speed optimization flags, I'm adding `/Ob2` for "inline more often", and `/arch:AVX2` in the expectation that it also enables `BMI2`. That's about it. Are there other settings that matter, and could influence measurements ?