@Niadb 2020-08-04
hi @Cyan4973 

That is about the same as what I saw, the compression speed diff was very small, basically irrelevant.

I was more interested in decompression speed, as that is what I mostly use zstd for.

1767.6/1727.9 is only about 2.2%, so in line with what you reported, it varies some each time I run fullbench, but is generally in the 2-3% range.

I think you basically had the same flags as me.  Intel vs AMD might make some difference, not sure what CPU you have.

For comparison here is clang on the same CPU as the above ^, it is very far ahead of MSVC in every metric

*** Zstandard speed analyzer 1.4.5 64-bits, by Yann Collet (Jul 28 2020) ***
Sample 10000000 bytes :
1#compress : 646.0 MB/s ( 3154550)
2#decompress : 2131.9 MB/s (10000000)
11#compressContinue : 659.2 MB/s ( 3154550)
12#compressContinue_extDict : 624.6 MB/s ( 3154887)
13#decompressContinue : 2021.5 MB/s (10000000)
31#decodeLiteralsBlock : 3057.5 MB/s ( 37031)
32#decodeSeqHeaders : 63629.0 MB/s ( 73)
41#compressStream : 478.1 MB/s ( 3154678)
42#decompressStream : 2046.9 MB/s (10000000)
43#compressStream_freshCCtx : 482.7 MB/s ( 3154678)
51#compress_generic, continue : 481.7 MB/s ( 3154678)
52#compress_generic, end : 643.3 MB/s ( 3154550)
61#compress_generic, -T2, contin: 486.5 MB/s ( 3154678)
62#compress_generic, -T2, end : 662.8 MB/s ( 3154550)